--- configs/example.yaml
@@ -1,59 +1,0 @@
-# ResNet18 - ~4.5M parameters
-resnet18:
-  model: "resnet18"
-  num_blocks: [2, 2, 2, 2]
-  channels_per_layer: [64, 128, 256, 512]
-  conv_kernel_sizes: [3, 3, 3, 3]
-  shortcut_kernel_sizes: [1, 1, 1, 1]
-  drop: 0.2
-  squeeze_and_excitation: false
-
-# ResNet18-Slim - ~1.7M parameters
-resnet18_slim:
-  model: "resnet18"
-  num_blocks: [2, 2, 2, 2]
-  channels_per_layer: [32, 64, 128, 256]
-  conv_kernel_sizes: [3, 3, 3, 3]
-  shortcut_kernel_sizes: [1, 1, 1, 1]
-  drop: 0.2
-  squeeze_and_excitation: false
-
-# ResNet20 - ~2.7M parameters
-resnet20:
-  model: "resnet20"
-  num_blocks: [3, 3, 3]
-  channels_per_layer: [48, 96, 192]
-  conv_kernel_sizes: [3, 3, 3]
-  shortcut_kernel_sizes: [1, 1, 1]
-  drop: 0.2
-  squeeze_and_excitation: false
-
-# ResNet20-SE - ~2.8M parameters
-resnet20_se:
-  model: "resnet20"
-  num_blocks: [3, 3, 3]
-  channels_per_layer: [48, 96, 192]
-  conv_kernel_sizes: [3, 3, 3]
-  shortcut_kernel_sizes: [1, 1, 1]
-  drop: 0.2
-  squeeze_and_excitation: true
-
-# ResNet34-MiniChannels - ~4.8M parameters
-resnet34_mini:
-  model: "resnet34"
-  num_blocks: [3, 4, 6, 3]
-  channels_per_layer: [32, 64, 128, 256]
-  conv_kernel_sizes: [3, 3, 3, 3]
-  shortcut_kernel_sizes: [1, 1, 1, 1]
-  drop: 0.1
-  squeeze_and_excitation: false
-
-# ResNet56-Micro - ~4.3M parameters
-resnet56_micro:
-  model: "resnet56"
-  num_blocks: [9, 9, 9]
-  channels_per_layer: [32, 64, 128]
-  conv_kernel_sizes: [3, 3, 3]
-  shortcut_kernel_sizes: [1, 1, 1]
-  drop: 0.1
-  squeeze_and_excitation: false
