+++ configs/experiments/DeepSlim.yaml
@@ -1,0 +1,26 @@
+# DeepSlim : 4.9M parameters
+# Deeper but thinner architecture with targeted regularization
+
+num_blocks: [3, 4, 4, 2]  
+conv_kernel_sizes: [3, 3, 3, 3]
+shortcut_kernel_sizes: [1, 1, 1, 1]
+num_channels: 48  
+avg_pool_kernel_size: 4  
+drop: 0.15  
+squeeze_and_excitation: true
+grad_clip: 0.1
+max_epochs: 200
+optim: "sgd"
+lr_sched: "CosineAnnealingLR"
+momentum: 0.9
+lr: 0.08
+weight_decay: 0.0005
+batch_size: 256
+num_workers: 16
+data_augmentation: 1
+data_normalize: 1
+lookahead: 0
+label_smoothing: 0.1
+use_amp: true
+early_stopping_patience: 12
+min_delta : 0.001
