+++ configs/experiments/ResReg.yaml
@@ -1,0 +1,24 @@
+# ResReg : 3.4M parameters
+# Uses varied kernel sizes and strong regularization
+
+num_blocks: [3, 4, 2]
+conv_kernel_sizes: [5, 3, 3]  # Larger kernels early
+shortcut_kernel_sizes: [1, 1, 1]
+num_channels: 56
+avg_pool_kernel_size: 8
+drop: 0.3  # Moderate dropout
+squeeze_and_excitation: true
+grad_clip: 0.05  # Tighter gradient clipping
+max_epochs: 120
+optim: "rmsprop"  # RMSProp optimizer
+lr_sched: "StepLR"  # Step learning rate
+momentum: 0.9
+lr: 0.001
+weight_decay: 0.0015  # Strong weight decay
+batch_size: 256
+num_workers: 16
+data_augmentation: 1
+data_normalize: 1
+lookahead: 1
+label_smoothing: 0.15
+use_amp: true
