+++ configs/experiments/ResWide.yaml
@@ -1,0 +1,26 @@
+# ResWide : 3.8M parameters
+# Wider but shallower network with targeted regularization
+
+num_blocks: [2, 2, 1]  # Reduced depth
+conv_kernel_sizes: [3, 3, 5]  # Larger kernels at deeper layers
+shortcut_kernel_sizes: [3, 1, 1]  # Wider shortcuts early
+num_channels: 80  # Wider channels
+avg_pool_kernel_size: 8
+drop: 0.25
+squeeze_and_excitation: true
+grad_clip: 0.1
+max_epochs: 180
+optim: "adamw"  # AdamW optimizer
+lr_sched: "OneCycleLR"  # One cycle learning rate
+momentum: 0.85
+lr: 0.003
+weight_decay: 0.001
+batch_size: 256
+num_workers: 16
+data_augmentation: 1
+data_normalize: 1
+lookahead: 1
+label_smoothing: 0.1
+use_amp: true
+early_stopping_patience: 12
+min_delta : 0.0009
